Adds Docker support with Docker Compose
Introduces Dockerfiles for the API and Blazor applications,
enabling containerization for easier deployment and scaling.

Includes a Docker Compose file for orchestrating the API and
Blazor services, simplifying local development and testing.

Adds a GitHub Actions workflow to build and push Docker images
to Docker Hub on pushes to the main branch, and on pull requests.

Also includes a PowerShell script for local building and pushing
of the docker images.